2.projects/sglang-inference/requirements.txt
aws-cdk-lib==2.177.0
constructs>=10.0.0,<11.0.0
PyYAML>=6.0